services/orchestrator/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.2
python-dotenv==1.0.1
chromadb==0.5.5
sentence-transformers==3.0.1
numpy==1.26.4
edge-tts==6.1.15
aiofiles==23.2.1
